隐私政策移动到服务器

+++ module_video/src/main/java/com/example/module_video/ui/widget/FloatPlayerView.kt
@@ -29,7 +29,7 @@ class FloatPlayerView @JvmOverloads constructor(
     }
 
 
-    fun setCurrentResource(list:List<MediaInformation>,position:Int,currentPositionWhenPlaying:Int){
+    fun setCurrentResource(list:List<MediaInformation>,position:Int,currentPositionWhenPlaying:Long){
         if (list.isNotEmpty()){
             videoPlayer.setUp(list, position,false, list[position].name)
             videoPlayer.seekOnStart=currentPositionWhenPlaying.toLong()
